Moved helper functions from misc to unpacking
